+++ tests/testthat/test_sanity.R
@@ -1,4 +1,5 @@
 test_that("sanity: required packages are available", {
   expect_true(requireNamespace("targets", quietly = TRUE))
+  expect_true(requireNamespace("parallel", quietly = TRUE))
   expect_true(requireNamespace("tarchetypes", quietly = TRUE))
 })